Fix server firmware start policy
After switching to a non-English language and bmc is taken over by
HMC, `Server firmware start policy` will not be displayed, because it
is hard-code, so only English is supported.

This commit is to use a variable to filter `Server firmware start
policy` instead of hardcoding.

Tested: Manually changed to non-English, built webui-vue successfully
and passes.

Signed-off-by: George Liu <liuxiwei@inspur.com>