Adding Led's to enhance use experience
- Adding custom LED component to enhance visual representation of
physical LED State which will be dipict the original system LED
colour and it's behaviour.
- When user toggle the identify LED button then as current behaviour
toast message will be appears for the LED status but as part of this
change when user toggle identify LED button then it's appears as LED
button and colour of the LED continuously blinking based on toggle
button status.
- For each toggle button behaviour is different as similar as Hardware
LED colour code.
- For example:
Dimms, fans, processor etc LED colour representation are amber and
system identify LED colour is blue as similar as Hardware colour.

Signed-off-by: Kamalkumar <108823339+Patel-Kamalkumar@users.noreply.github.com>